Move to `FileChooserNative`
Enables Flatpak users to see the dialog of their desktop environment.

Changes the responses to the ones in this dialog which are
REPONSE_ACCEPT and RESPONSE_CANCEL.

It also cleans up the code for the suggested action that is in the hands
of the platform. The GTK/GNOME file selector has open as the suggested
action, for example.